# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.SwaggerSimpleResponse:
    properties:
      message:
        type: string
      status:
        type: boolean
    type: object
  api.SwaggerWebhookConfigListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.WebhookConfigResponse'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  api.SwaggerWebhookConfigResponse:
    properties:
      data:
        $ref: '#/definitions/dto.WebhookConfigResponse'
      message:
        type: string
      status:
        type: boolean
    type: object
...
  dto.WebhookConfigRequest:
    properties:
      paye_webhook_slug:
        type: string
      provider_name:
        type: string
      target_url:
        type: string
    required:
    - provider_name
    - target_url
    type: object
  dto.WebhookConfigResponse:
    properties:
      id:
        type: string
      paye_webhook_slug:
        type: string
      provider_name:
        type: string
      target_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Unified payment infrastructure API.
  title: Paye API
  version: "1.0"
paths:
...
  /webhooks/configs:
    get:
      description: Retrieve all webhook configs for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwaggerWebhookConfigListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
      security:
      - BearerAuth: []
      summary: List webhook configurations
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Register a webhook callback target for a specific provider
      parameters:
      - description: Webhook Configuration Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WebhookConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwaggerWebhookConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
      security:
      - BearerAuth: []
      summary: Create webhook configuration
      tags:
      - Webhooks
  /webhooks/configs/{id}:
    delete:
      description: Remove a specific webhook config
      parameters:
      - description: Webhook Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
      security:
      - BearerAuth: []
      summary: Delete webhook configuration
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Modify a specific webhook config
      parameters:
      - description: Webhook Config ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook Config Update Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WebhookConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SwaggerWebhookConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
      security:
      - BearerAuth: []
      summary: Update webhook configuration
      tags:
      - Webhooks
  /webhooks/receive/{slug}:
    post:
      consumes:
      - application/json
      description: Public proxy endpoint that receives webhooks from providers (e.g.
        Paystack) and forwards them to target URL
      parameters:
      - description: Webhook unique slug
        in: path
        name: slug
        required: true
        type: string
      - description: Paystack signature for verification
        in: header
        name: X-Paystack-Signature
        type: string
      - description: Provider Webhook payload
        in: body
        name: request
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.SwaggerSimpleResponse'
      summary: Receive payment provider webhook
      tags:
      - Webhooks Public
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"